fixed integer types

+++ db/migrate/20231028013139_create_transactions.rb
@@ -3,13 +3,13 @@
 class CreateTransactions < ActiveRecord::Migration[7.1]
   def change
     create_table :transactions do |t|
-      t.integer :transaction_id
-      t.integer :merchant_id
-      t.integer :user_id
+      t.bigint :transaction_id
+      t.bigint :merchant_id
+      t.bigint :user_id
       t.string :card_number
       t.datetime :transaction_date
       t.decimal :transaction_amount
-      t.integer :device_id
+      t.bigint :device_id
       t.boolean :has_cbk
 
       t.timestamps
